Get a clean `-fsanitize=integer` run of `make check`

+++ src/reduced_basis/rb_construction.C
@@ -269,7 +269,7 @@ void RBConstruction::process_parameters_file (const std::string & parameters_fil
   // Set the parameters that have been read in
   set_rb_construction_parameters(n_training_samples,
                                  deterministic_training,
-                                 training_parameters_random_seed_in,
+                                 static_cast<int>(training_parameters_random_seed_in),
                                  quiet_mode_in,
                                  Nmax_in,
                                  rel_training_tolerance_in,
@@ -285,7 +285,7 @@ void RBConstruction::process_parameters_file (const std::string & parameters_fil
 void RBConstruction::set_rb_construction_parameters(
                                                     unsigned int n_training_samples_in,
                                                     bool deterministic_training_in,
-                                                    unsigned int training_parameters_random_seed_in,
+                                                    int training_parameters_random_seed_in,
                                                     bool quiet_mode_in,
                                                     unsigned int Nmax_in,
                                                     Real rel_training_tolerance_in,
